added unit testing

+++ greenfield.jms/src/main/java/greenfield/jms/DomainEvent.java
@@ -2,14 +2,14 @@ package greenfield.jms;
 
 public class DomainEvent {
 
-	public Long id;
+	public Integer id;
 	public String value;
 
 	public DomainEvent() {
 
 	}
 
-	public DomainEvent(Long id, String value) {
+	public DomainEvent(Integer id, String value) {
 		super();
 		this.id = id;
 		this.value = value;
